Move serialization tests inside the state machine.
Users will have a clean DB, so this is the better place to test.
