+++ raiden/transfer/architecture.py
@@ -9,6 +9,7 @@ from raiden.utils.typing import (
     T_ChannelID,
     TransactionHash,
 )
+from raiden.storage.serialize import JSONSerializer
 from raiden.utils.notifying_queue import QueueIdentifier
 
 
@@ -35,6 +36,55 @@ from raiden.utils.notifying_queue import QueueIdentifier
 # processed, i.e. the state change must be self contained and the result state
 # tree must be serializable to produce a snapshot. To enforce this inputs and
 # outputs are separated under different class hierarquies (StateChange and Event).
+
+
+def compare_state_trees(state1, state2):
+    print('> Looking at', state1.__class__.__name__)
+    if not state1.__class__.__name__ == state2.__class__.__name__:
+        print(
+            'Instances do not have same type:',
+            state1.__class__.__name__,
+            state2.__class__.__name__,
+        )
+        return
+
+    if isinstance(state1, dict):
+        if state1.keys() != state2.keys():
+            print('dics do not have same keys:', state1.keys(), state2.keys())
+        else:
+            for k in state1.keys():
+                compare_state_trees(state1[k], state2[k])
+    elif isinstance(state1, list):
+        if len(state1) != len(state2):
+            print('lists do not have same length:', len(state1), len(state2))
+        else:
+            for i in range(len(state1)):
+                compare_state_trees(state1[i], state2[i])
+    else:
+        slots1 = state1.__slots__
+        slots2 = state2.__slots__
+
+        if not slots1 == slots2:
+            print('Instances do not have same slots:', slots1, slots2)
+            return
+
+        for attr_name in state1.__slots__:
+            if attr_name in ('pseudo_random_generator'):
+                return
+
+            print('> Comparing', attr_name)
+
+            attr1 = getattr(state1, attr_name)
+            attr2 = getattr(state2, attr_name)
+
+            if not attr1 == attr2:
+                from raiden.transfer.state import State
+                print('Instances do not have same value:', attr1, attr2)
+                if isinstance(attr1, (State, dict, list)):
+                    print('... Recurse into compare_state_trees')
+                    compare_state_trees(attr1, attr2)
+            else:
+                print('>>> Matching')
 
 
 class State:
@@ -195,6 +245,18 @@ class StateManager:
         """
         assert isinstance(state_change, StateChange)
 
+        # check state change serialization
+        try:
+            json = JSONSerializer.serialize(state_change)
+            restored = JSONSerializer.deserialize(json)
+
+            if state_change != restored:
+                print('Serialisation failed for:', state_change.__class__.__name__)
+            else:
+                print('State change serialisation round-trip successful')
+        except Exception:
+            print('Serialisation failed for:', state_change.__class__.__name__)
+
         # the state objects must be treated as immutable, so make a copy of the
         # current state and pass the copy to the state machine to be modified.
         next_state = deepcopy(self.current_state)
@@ -209,6 +271,18 @@ class StateManager:
 
         self.current_state = iteration.new_state
         events = iteration.events
+
+        # check state serialization
+        state = self.current_state
+        if state is not None:
+            json = JSONSerializer.serialize(state)
+            restored = JSONSerializer.deserialize(json)
+
+            if state != restored:
+                print('###########################################')
+                compare_state_trees(state, restored)
+            else:
+                print('State serialisation round-trip successful')
 
         assert isinstance(self.current_state, (State, type(None)))
         assert all(isinstance(e, Event) for e in events)
